backend/requirements.txt
fastapi
uvicorn
python-pptx
pdfminer.six
pydantic
python-multipart